Rename sampler_integrator
